[C#] refactor: use fluent api in chef and gpt me samples
